+++ packages/components/src/Tooltip/Tooltip.tsx
@@ -36,7 +36,7 @@ import React, {
   ReactNode,
   Ref,
 } from 'react'
-import { omit } from 'lodash'
+import omit from 'lodash/omit'
 import { ModalContext } from '../Modal'
 import {
   useCallbackRef,