+++ CHANGELOG.md
@@ -11,7 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Tree`, `TreeItem` components
 
-## [0.7.32] - 2020-05-19
+## [0.7.37] - 2020-05-20
+
+### Fixed
+
+- `Space` & `SpaceVertical` default to 100% width of their container again
+
+## [0.7.36] - 2020-05-20
 
 ### Added
 
@@ -29,6 +35,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - now supports `around`, `between` and `evenly` as alternatives to `gap`
   - now supports `verticalAlign` property
 - `SpaceVertical` now supports `align` and `stretch` properties
+- `Accordion`, `AccordionLabel`, `AccordionContent` components
+- `ButtonToggle` now accepts `nullable`
+- `ButtonGroup` and `ButtonToggle` will now wrap if there are too many items for the container width
+- created `FieldRangeSlider` and `FieldSlider` components
 
 ### Changed